mts_hw_module/setup/requirements.txt
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
blinker==1.5
Brotli==1.0.9
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_83242e7e-f82d-4a71-8ef2-9d71d212d249gu_wxmeq/croots/recipe/certifi_1655968827803/work/certifi
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
fake-useragent==0.1.11
greenlet==1.1.2
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.3
kaitaistruct==0.10
lxml==4.9.1
outcome==1.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pyTelegramBotAPI==4.6.0
python-dotenv==0.20.0
requests==2.28.1
selenium==4.3.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.10
webdriver-manager==3.8.2
wsproto==1.1.0
zstandard==0.18.0
